+++ pkg/persistence/k8s_envs.go
@@ -1,6 +1,7 @@
 package persistence
 
 import (
+	"context"
 	"database/sql"
 
 	"github.com/dollarshaveclub/acyl/pkg/models"
@@ -10,10 +11,13 @@ import (
 var _ K8sEnvDataLayer = &PGLayer{}
 
 // GetK8sEnv gets a k8s environment by environment name
-func (pg *PGLayer) GetK8sEnv(name string) (*models.KubernetesEnvironment, error) {
+func (pg *PGLayer) GetK8sEnv(ctx context.Context, name string) (*models.KubernetesEnvironment, error) {
 	out := &models.KubernetesEnvironment{}
+	if isCancelled(ctx.Done()) {
+		return nil, errors.Wrap(ctx.Err(), "error getting k8s env")
+	}
 	q := `SELECT ` + models.KubernetesEnvironment{}.Columns() + ` FROM kubernetes_environments WHERE env_name = $1;`
-	if err := pg.db.QueryRow(q, name).Scan(out.ScanValues()...); err != nil {
+	if err := pg.db.QueryRowContext(ctx, q, name).Scan(out.ScanValues()...); err != nil {
 		if err == sql.ErrNoRows {
 			return nil, nil
 		}
@@ -23,29 +27,41 @@ func (pg *PGLayer) GetK8sEnv(name string) (*models.KubernetesEnvironment, error)
 }
 
 // GetK8sEnvsByNamespace returns any KubernetesEnvironments associated with a specific namespace name
-func (pg *PGLayer) GetK8sEnvsByNamespace(ns string) ([]models.KubernetesEnvironment, error) {
+func (pg *PGLayer) GetK8sEnvsByNamespace(ctx context.Context, ns string) ([]models.KubernetesEnvironment, error) {
+	if isCancelled(ctx.Done()) {
+		return nil, errors.Wrap(ctx.Err(), "error getting k8s env by namespace")
+	}
 	q := `SELECT ` + models.KubernetesEnvironment{}.Columns() + ` FROM kubernetes_environments WHERE namespace = $1;`
-	return collectK8sEnvRows(pg.db.Query(q, ns))
+	return collectK8sEnvRows(pg.db.QueryContext(ctx, q, ns))
 }
 
 // CreateK8sEnv inserts a new k8s environment into the DB
-func (pg *PGLayer) CreateK8sEnv(env *models.KubernetesEnvironment) error {
+func (pg *PGLayer) CreateK8sEnv(ctx context.Context, env *models.KubernetesEnvironment) error {
+	if isCancelled(ctx.Done()) {
+		return errors.Wrap(ctx.Err(), "error creating k8s env")
+	}
 	q := `INSERT INTO kubernetes_environments (` + env.InsertColumns() + `) VALUES (` + env.InsertParams() + `);`
-	_, err := pg.db.Exec(q, env.InsertValues()...)
+	_, err := pg.db.ExecContext(ctx, q, env.InsertValues()...)
 	return errors.Wrap(err, "error inserting k8s environment")
 }
 
 // DeleteK8sEnv deletes a k8s environment from the DB
-func (pg *PGLayer) DeleteK8sEnv(name string) error {
+func (pg *PGLayer) DeleteK8sEnv(ctx context.Context, name string) error {
+	if isCancelled(ctx.Done()) {
+		return errors.Wrap(ctx.Err(), "error deleting k8s env")
+	}
 	q := `DELETE FROM kubernetes_environments WHERE env_name = $1;`
-	_, err := pg.db.Exec(q, name)
+	_, err := pg.db.ExecContext(ctx, q, name)
 	return errors.Wrap(err, "error deleting k8s environment")
 }
 
 // UpdateK8sEnvTillerAddr updates an existing k8s environment with the provided tiller address
-func (pg *PGLayer) UpdateK8sEnvTillerAddr(envname, taddr string) error {
+func (pg *PGLayer) UpdateK8sEnvTillerAddr(ctx context.Context, envname, taddr string) error {
+	if isCancelled(ctx.Done()) {
+		return errors.Wrap(ctx.Err(), "error update k8s env tiller address")
+	}
 	q := `UPDATE kubernetes_environments SET tiller_addr = $1 WHERE env_name = $2;`
-	_, err := pg.db.Exec(q, taddr, envname)
+	_, err := pg.db.ExecContext(ctx, q, taddr, envname)
 	return errors.Wrap(err, "error updating k8s environment")
 }
